Extractive_Text_Summarization/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.1
bleach==6.0.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
cssselect==1.2.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.16.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.0
Flask==2.0.2
Flask-Cors==3.0.10
fqdn==1.5.1
gunicorn==20.1.0
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==4.8.1
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.18.2
jieba3k==0.35.1
Jinja2==3.0.2
joblib==1.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
murmurhash==1.0.9
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
notebook_shim==0.2.2
numpy==1.24.3
packaging==23.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2021.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.0.1
scipy==1.10.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
thinc==8.1.9
threadpoolctl==3.0.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.3
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typer==0.7.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.0.2
zipp==3.6.0